Corrected the link args
The linker wasn't able to find libraries that skia-bindings was binding
to. Prior to this patch, I was pointing the linker to these libraries by
directing the build to the .so full path.

This patch allows the linker to determine the path which should
hopefully work for systems where libs may be in a different place (or a
lib updates to a newer version and the .so file changes names).

It's possible to bring in pkg-config to help with this but I feel like
that isn't necessary in this scenario. This would mean a massive
modification of the build.rs in comparison to passing linker args which
is a few lines for this patch to work.