Removed unnessary neovide build depends
On the hopefully final tweak to the neovide ebuild, I have removed the
BDEPENDS which were needed to build skia + bindings. This was namely
clang, ninja, gn and pkg-config.

Since the binaries also don't need the source code to skia +
depot_tools, the build doesn't download those. This should save
bandwidth and unnessary disk usage.

I could potentially in the future offer an IUSE to build skia but this
will likely be something I work on in the distant future.