neovide: 0.13.1 bump, 0.12.1 & 9999 removal
I'm a couple of weeks late but 0.13.1 is here. 0.13.0 brought a pretty
big breaking change which called for neovim-0.10.0 as the base minimum.
At the moment, that's marked as ~amd64 so be sure to unmask it. Not
doing so will still allow you to install neovide 0.13.1 but you'll
encounter an immediate neovide crash.

Now about the removals: 0.12.1 has been removed since 0.13.1 is the
newest version. I've also removed the eternally broken 9999 ebuild.

I've also removed all of the IUSE flags as they weren't exactly doing
anything. Users building this are likely running X or Wayland. Lunarvim
files have also been removed as I honestly can't be bothered maintaining
the patches anymore.

Signed-off-by: nosqrt <nosqrt@outlook.com>