app-editors/neovide: Add arm64 build
Since I have a raspberry pi 5, I am able to finally test neovide on
arm64 hardware. The big difference with the ebuild is the download
and handling of the prebuilt skia-binaries packages.

I also added a pkg-postinst message explaining a work around for opengl
not working for some machines. The Pi 5 is capable of opengl 3.1 while
the neovide project uses opengl 3.3. Luckily there is a way to override
the opengl context failure by using an environment variable to instead
use the CPU to handle rendering.

Signed-off-by: Daniella Kicsak <nosqrt@outlook.com>