app-editor/neovide: Cleanup and update to 0.14.0
Most of the old builds likely no longer work due to some changes
regarding cargo and rust work within gentoo.

The 0.14.0 ebuild now has shifted some of the linking to the ebuild
itself instead of the build.rs patch which was doing that heavy
lifting in the past. GL and EGL don't seem to work using the export
method which is likely due to using a binary version of skia. When I
have more motivation and time, I'll investigate this further.

Signed-off-by: Daniella Kicsak <nosqrt@outlook.com>