dev-python/mdns-publish-cname: Add openrc files
Wrote up a quick init script so that openrc itself manages the daemon.
There's alternatives ways to launch it (e.g. cron) but this'll do for
now. A systemd .ini could also be added but I don't have much use for it
atm as I'm not running this on my main desktop.