dev-perl/DBD-Pg: Forked to use c17. c23 mitigation
Because the DBD-Pg code-base hasn't yet updated the code to work with
the c23 specification (default in GCC 15), I've added the "quick and
dirty" hack of specifying to use the c17 protocol. This isn't a great
fix but I like being able to talk to postgres from perl and don't
particularly feel like deep-diving into the code that generates the
header table.

While I was at it, I created a patch to update to github master head for
some of my more recent fixes in the past couple of years.

Signed-off-by: Daniella Kicsak <nosqrt@outlook.com>